11-12

+++ backend/controllers/CompetController.php
@@ -9,6 +9,7 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\UploadedFile;
+use yii\web\ForbiddenHttpException;
 
 /**
  * CompetController implements the CRUD actions for Compet model.
@@ -73,6 +74,9 @@ class CompetController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::$app->user->can('createCompet')){
+            throw new ForbiddenHttpException('对不起您没有进行该操作的权限');
+        }
         $model = new Compet();
 
         if($model->load(Yii::$app->request->post())){
@@ -97,6 +101,9 @@ class CompetController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(!Yii::$app->user->can('updateCompet')){
+            throw new ForbiddenHttpException('对不起您没有进行该操作的权限');
+        }
         $model = $this->findModel($id);
 
         if($model->load(Yii::$app->request->post())){
@@ -120,6 +127,9 @@ class CompetController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->can('deleteCompet')){
+            throw new ForbiddenHttpException('对不起您没有进行该操作的权限');
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
