+++ backend/controllers/OldbookController.php
@@ -9,6 +9,7 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\UploadedFile;
+use yii\web\ForbiddenHttpException;
 
 /**
  * OldbookController implements the CRUD actions for Oldbook model.
@@ -73,6 +74,9 @@ class OldbookController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::$app->user->can('createOldbook')){
+            throw new ForbiddenHttpException('对不起您没有进行该操作的权限');
+        }
         $model = new Oldbook();
 
         if($model->load(Yii::$app->request->post())){
@@ -97,6 +101,9 @@ class OldbookController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(!Yii::$app->user->can('updateOldbook')){
+            throw new ForbiddenHttpException('对不起您没有进行该操作的权限');
+        }
         $model = $this->findModel($id);
         if($model->load(Yii::$app->request->post())){
             $model=self::upload($model);
@@ -119,6 +126,9 @@ class OldbookController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->can('deleteOldbook')){
+            throw new ForbiddenHttpException('对不起您没有进行该操作的权限');
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
